Fix `FIAT 4200`
While testing shutter on optimism, we saw `4200` wei weirdly appearing
on the receiver of the encrypted tx. This came from the
`excessGasRecipient` in https://github.com/shutter-network/shop-contracts/blob/main/src/Inbox.sol#L68 being set to the receiver address. Sending the gas
back to the sender is the more expected behavior, so we change the
parameter to `tx.from`.